Minor renames in Chat API
