+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/ChatContainer.kt
@@ -60,7 +60,7 @@ internal class ChatContainer(
             started = true
             configuration.chatConfig =
                 ChatConfiguration(
-                    endPointURL = remoteOptions.endpointUrl,
+                    endpoint = remoteOptions.endpoint,
                     identity = remoteOptions.identity,
                     credential = remoteOptions.credential,
                     applicationID = DiagnosticConfig().tag,
