+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/models/ChatCompositeException.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.ui.chat;
+package com.azure.android.communication.ui.chat.models;
 
 
 /**
